requirements.txt
# langchain
# langchain-community
# langchain-core
# openai
# pypdf
tiktoken
# aiofiles
# fastapi
# uvicorn
jinja2
# python-multipart
# PyPDF2
# faiss-cpu
# -e .

langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
fastapi==0.104.1
uvicorn==0.24.0
aiofiles==23.2.1
python-dotenv==1.0.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
-e .
